+++ src/common/util.cpp
@@ -23,6 +23,7 @@
 #include <wex/syntax/lexers.h>
 #include <wex/syntax/stc.h>
 #include <wx/app.h>
+#include <wx/choicdlg.h>
 #include <wx/wupdlock.h>
 
 namespace wex
@@ -187,8 +188,10 @@ bool wex::lexers_dialog(syntax::stc* stc)
       return i.display_lexer();
     });
 
-  if (auto lexer = stc->get_lexer().display_lexer();
-      !single_choice_dialog(stc, _("Enter Lexer"), s, lexer))
+  if (auto lexer = stc->get_lexer().display_lexer(); !single_choice_dialog(
+        data::window().parent(stc).title(_("Enter Lexer")),
+        s,
+        lexer))
   {
     return false;
   }
@@ -277,6 +280,35 @@ bool wex::shell_expansion(std::string& command)
       r.replace(command, process.std_out());
     }
   }
+
+  return true;
+}
+
+bool wex::single_choice_dialog(
+  const data::window&             data,
+  const std::vector<std::string>& v,
+  std::string&                    selection)
+{
+  wxArrayString s;
+
+  for (const auto& it : v)
+  {
+    s.Add(it);
+  }
+
+  wxSingleChoiceDialog dlg(data.parent(), _("Input") + ":", data.title(), s);
+
+  if (data.size() != wxDefaultSize)
+  {
+    dlg.SetSize(data.size());
+  }
+
+  if (const auto index = s.Index(selection); index != wxNOT_FOUND)
+    dlg.SetSelection(index);
+  if (dlg.ShowModal() == wxID_CANCEL)
+    return false;
+
+  selection = dlg.GetStringSelection();
 
   return true;
 }