+++ src/ui/item-create-window.cpp
@@ -254,7 +254,7 @@ void create_font_picker_control(
     parent,
     item.data().window().id(),
     wxNullFont,
-    PSS == data::NUMBER_NOT_SET ? wxFNTP_DEFAULT_STYLE :
+    PSS == data::NUMBER_NOT_SET ? wxFNTP_DEFAULT_STYLE | wxPB_SMALL :
                                   item.data().window().style());
 
   finish_picker(pc, item, window);